fix(iframeAPI): startShareVideo command.
