+++ react/features/shared-video/components/AbstractSharedVideoDialog.js
@@ -3,7 +3,7 @@
 import { Component } from 'react';
 import type { Dispatch } from 'redux';
 
-import { getYoutubeId } from '../functions';
+import { extractYoutubeIdOrURL } from '../functions';
 
 /**
  * The type of the React {@code Component} props of
@@ -56,34 +56,15 @@ export default class AbstractSharedVideoDialog<S: *> extends Component < Props, 
      * @returns {boolean}
      */
     _onSetVideoLink(link: string) {
-        if (!link) {
-            return false;
-        }
-
-        const trimmedLink = link.trim();
-
-        if (!trimmedLink) {
-            return false;
-        }
-
         const { onPostSubmit } = this.props;
-        const youtubeId = getYoutubeId(trimmedLink);
 
-        if (youtubeId) {
-            onPostSubmit(youtubeId);
+        const id = extractYoutubeIdOrURL(link);
 
-            return true;
-        }
-
-        // Check if the URL is valid, native may crash otherwise.
-        try {
-            // eslint-disable-next-line no-new
-            new URL(trimmedLink);
-        } catch (_) {
+        if (!id) {
             return false;
         }
 
-        onPostSubmit(trimmedLink);
+        onPostSubmit(id);
 
         return true;
     }
